feat: extract form items to sigle files
