+++ src/renderer/components/setting/settings-form/avatar-uploader.tsx
@@ -2,8 +2,8 @@ import { convertFileToBase64 } from '@/lib/utils'
 import { AlertCircle, Edit, Trash } from 'lucide-react'
 import { ChangeEvent, useRef } from 'react'
 import { FieldValues, useController, UseControllerProps } from 'react-hook-form'
-import { Alert, AlertDescription } from './ui/alert'
-import { Button } from './ui/button'
+import { Alert, AlertDescription } from '../../ui/alert'
+import { Button } from '../../ui/button'
 
 export function AvatarUploader<T extends FieldValues>({
   props
