+++ src/renderer/components/setting/settings-form/providers/deepseek.tsx
@@ -1,0 +1,54 @@
+import {
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage
+} from '@/components/ui/form'
+import { Input } from '@/components/ui/input'
+import { UseFormReturnType } from '../../settings-form'
+
+export function DeepSeek({ form }: { form: UseFormReturnType }) {
+  return (
+    <>
+      <FormField
+        control={form.control}
+        name="deepSeekApiKey"
+        render={({ field }) => (
+          <FormItem>
+            <FormLabel>API Key</FormLabel>
+            <FormControl>
+              <Input
+                type="password"
+                autoComplete="current-password"
+                id="deep-seek-api-key-input"
+                autoFocus
+                {...field}
+                value={field.value ?? ''}
+              />
+            </FormControl>
+            <FormMessage />
+          </FormItem>
+        )}
+      />
+      <FormField
+        control={form.control}
+        name="deepSeekBaseUrl"
+        render={({ field }) => (
+          <FormItem>
+            <FormLabel>Base URL</FormLabel>
+            <FormControl>
+              <Input
+                type="text"
+                id="deep-seek-base-url-input"
+                {...field}
+                value={field.value ?? ''}
+              />
+            </FormControl>
+            <FormMessage />
+          </FormItem>
+        )}
+      />
+    </>
+  )
+}
